Refactor SystemTestMixin into its own module.
